Add option to log which jail each banned IP came from
When banned IPs are parsed from fail2ban jails and added to the
blacklist, the jail that caused the ban can be added as an inline comment
in the blacklist file.

Jail logging is disabled by default. Enable by setting LOG_JAILS=true in
the conf, or calling the script with the -l option.

When the same IP appears multiple times from different jails, the first
occurance is kept as determined by sort -u, which by default is
alphabetical.